Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée

Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée

Utilisation

cartofriches_friches(
  coddep = NULL,
  code_insee = NULL,
  contains_geom = NULL,
  fields = NULL,
  in_bbox = NULL,
  ordering = NULL,
  surface_max = NULL,
  surface_min = NULL,
  urba_zone_type = NULL
)

Arguments

coddep

Code INSEE départemental

code_insee

Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)

contains_geom

Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom='type':'Point', 'coordinates':[2.17,46.75]

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

in_bbox

Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 1.0 deg. x 1.0 deg.

ordering

Which field to use when ordering the results.

surface_max

Surface maximale de l'unité foncière

surface_min

Surface minimale de l'unité foncière

urba_zone_type

Type de zone d'urbanisme

Valeur de retour

Retourne les friches issues de Cartofriches pour la commune, le département ou l'emprise rectangulaire demandée (paramètre code_insee, coddep ou in_bbox obligatoire)

Exemples

cartofriches_friches(coddep = 59)
#> NULL
cartofriches_friches(code_insee = 59350)
#> [1] 59350
#> # A tibble: 135 × 15
#>    site_id site_nom site_type site_adresse site_statut comm_nom comm_insee dep  
#>    <chr>   <chr>    <chr>     <lgl>        <chr>       <chr>    <chr>      <chr>
#>  1 59350_… Teintur… inconnu   NA           friche pot… LILLE    59350      59   
#>  2 59350_… WILVIA … inconnu   NA           friche pot… LILLE    59350      59   
#>  3 59350_… Atelier… inconnu   NA           friche pot… LILLE    59350      59   
#>  4 59350_… Station… inconnu   NA           friche pot… LILLE    59350      59   
#>  5 59350_… Assitan… inconnu   NA           friche pot… LILLE    59350      59   
#>  6 59350_… Constru… inconnu   NA           inconnu     LILLE    59350      59   
#>  7 59350_… Atelier… inconnu   NA           friche pot… LILLE    59350      59   
#>  8 59350_… Appel à… inconnu   NA           friche ave… LILLE    59350      59   
#>  9 59350_… Cie Eur… inconnu   NA           inconnu     LILLE    59350      59   
#> 10 59350_… Appel à… inconnu   NA           friche ave… LILLE    59350      59   
#> # ℹ 125 more rows
#> # ℹ 7 more variables: proprio_personne <chr>, unite_fonciere_surface <dbl>,
#> #   unite_fonciere_refcad <chr>, source_nom <chr>, proprio_type_lib <list>,
#> #   source_nature <chr>, urba_zone_type <chr>
cartofriches_friches(in_bbox = '2.827642,50.495958,3.32616,50.767734')
#> NULL