Mutations

Mutations

Utilisation

dv3f_mutations(
  token = NULL,
  anneemut = NULL,
  anneemut_max = NULL,
  anneemut_min = NULL,
  code_insee = NULL,
  codtypbien = NULL,
  codtypproa = NULL,
  codtypprov = NULL,
  contains_geom = NULL,
  fields = NULL,
  filtre = NULL,
  idnatmut = NULL,
  in_bbox = NULL,
  sbati_max = NULL,
  sbati_min = NULL,
  segmtab = NULL,
  sterr_max = NULL,
  sterr_min = NULL,
  valeurfonc_max = NULL,
  valeurfonc_min = NULL,
  vefa = NULL
)

Arguments

token

token

anneemut

Annee de la mutation (>= 2010) - cf anneemut

anneemut_max

Annee de la mutation maximale - cf anneemut

anneemut_min

Annee de la mutation minimale - cf anneemut

code_insee

Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)

codtypbien

Code(s) de la typologie de bien à sélectionner (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypbien=11,121 - cf codtypbien

codtypproa

Code(s) de la typologie de l'acheteur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypproa=P,F7 - cf codtypproa

codtypprov

Code(s) de la typologie du vendeur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypprov=P,F7 - cf codtypprov

contains_geom

Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom='type':'Point', 'coordinates':[2.17,46.75]

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

filtre

Code alphanumerique permettant d'exclure des transactions particulières - cf filtre

idnatmut

Code(s) de nature de mutation (il est possible d'en demander plusieurs en séparant par une virgule - exemple : idnatmut=1,2 - cf idnatmut

in_bbox

Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.

sbati_max

Surface maximale bâtie vendue - cf sbati

sbati_min

Surface minimale bâtie vendue - cf sbati

segmtab

Note de segment du terrain à bâtir - cf segmtab

sterr_max

Surface maximale de terrain vendue - cf sterr

sterr_min

Surface minimale de terrain vendue - cf sterr

valeurfonc_max

Valeur foncière maximale (€) - cf valeurfonc

valeurfonc_min

Valeur foncière minimale (€) - cf valeurfonc

vefa

Vente en l'état futur d'achèvement - cf vefa

Valeur de retour

Retourne, en GeoJSON, les tup issues des Fichiers Fonciers pour la commune ou l'emprise demandée (paramètre code_insee ou in_bbox obligatoire)

Exemples

if (FALSE) {
dv3f_mutations(code_insee=59001,token=token)
}