Géomutations open

Géomutations open

Utilisation

dvf_opendata_geomutations(
  anneemut = NULL,
  anneemut_max = NULL,
  anneemut_min = NULL,
  code_insee = NULL,
  codtypbien = NULL,
  codtypproa = NULL,
  codtypprov = NULL,
  contains_geom = NULL,
  fields = NULL,
  filtre = NULL,
  idnatmut = NULL,
  in_bbox = NULL,
  sbati_max = NULL,
  sbati_min = NULL,
  segmtab = NULL,
  sterr_max = NULL,
  sterr_min = NULL,
  valeurfonc_max = NULL,
  valeurfonc_min = NULL,
  vefa = NULL
)

Arguments

anneemut

Annee de la mutation (>= 2010) - cf anneemut

anneemut_max

Annee de la mutation maximale - cf anneemut

anneemut_min

Annee de la mutation minimale - cf anneemut

code_insee

Code INSEE communal ou d'arrondissement municipal (il est possible d'en demander plusieurs (10 maximum), séparés par des virgules, dans le même département)

codtypbien

Code(s) de la typologie de bien à sélectionner (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypbien=11,121 - cf codtypbien

codtypproa

Code(s) de la typologie de l'acheteur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypproa=P,F7 - cf codtypproa

codtypprov

Code(s) de la typologie du vendeur (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : codtypprov=P,F7 - cf codtypprov

contains_geom

Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom='type':'Point', 'coordinates':[2.17,46.75]

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

filtre

Code alphanumerique permettant d'exclure des transactions particulières - cf filtre

idnatmut

Code(s) de nature de mutation (il est possible d'en demander plusieurs en séparant par une virgule - exemple : idnatmut=1,2 - cf idnatmut

in_bbox

Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.

sbati_max

Surface maximale bâtie vendue - cf sbati

sbati_min

Surface minimale bâtie vendue - cf sbati

segmtab

Note de segment du terrain à bâtir - cf segmtab

sterr_max

Surface maximale de terrain vendue - cf sterr

sterr_min

Surface minimale de terrain vendue - cf sterr

valeurfonc_max

Valeur foncière maximale (€) - cf valeurfonc

valeurfonc_min

Valeur foncière minimale (€) - cf valeurfonc

vefa

Vente en l'état futur d'achèvement - cf vefa

Valeur de retour

Retourne, en GeoJSON, les mutations issues de DVF+ pour la commune ou l'emprise rectangulaire demandée (paramètre code_insee ou in_bbox obligatoire)

Exemples

dvf_opendata_geomutations(code_insee=59001)
#> No encoding supplied: defaulting to UTF-8.
#> Warning: GDAL Message 6: driver GeoJSON does not support open option ENCODING
#> [1] "84 / 84"
#> Simple feature collection with 84 features and 21 fields (with 1 geometry empty)
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: 3.174165 ymin: 50.15932 xmax: 3.305036 ymax: 50.25702
#> Geodetic CRS:  WGS 84
#> # A tibble: 84 × 22
#> # Groups:   idopendata [84]
#>    idmutinvar   idopendata datemut    anneemut coddep libnatmut vefa  valeurfonc
#>  * <chr>        <chr>      <date>        <int> <chr>  <chr>     <lgl> <chr>     
#>  1 274db1bcf69… 274db1bcf… 2022-08-05     2022 59     Vente     FALSE 47000.00  
#>  2 b05d9e0e984… b05d9e0e9… 2022-03-16     2022 59     Vente     FALSE 280000.00 
#>  3 950ab57f428… 950ab57f4… 2022-09-16     2022 59     Vente     FALSE 182100.00 
#>  4 7f3aaf2d727… 7f3aaf2d7… 2022-10-25     2022 59     Vente     FALSE 70000.00  
#>  5 7fca9a8e78f… 7fca9a8e7… 2022-10-26     2022 59     Vente     FALSE 280000.00 
#>  6 6faba458b70… 6faba458b… 2022-03-04     2022 59     Vente     FALSE 276000.00 
#>  7 5b670648771… 5b6706487… 2021-05-17     2021 59     Vente     FALSE 309600.00 
#>  8 06874211e78… 06874211e… 2021-12-28     2021 59     Vente     FALSE 8500.00   
#>  9 5e5f60095de… 5e5f60095… 2021-03-03     2021 59     Vente     FALSE 289000.00 
#> 10 2d561a1e58b… 2d561a1e5… 2021-02-26     2021 59     Vente     FALSE 225000.00 
#> # ℹ 74 more rows
#> # ℹ 14 more variables: nbcomm <int>, nbpar <int>, nbparmut <int>, sterr <chr>,
#> #   nbvolmut <int>, nblocmut <int>, sbati <chr>, codtypbien <chr>,
#> #   libtypbien <chr>, l_codinsee <chr>, l_idpar <chr>, l_idparmut <chr>,
#> #   l_idlocmut <chr>, geometry <POLYGON [°]>