GeoTUP

GeoTUP

Utilisation

ff_geotups(
  token,
  code_insee = NULL,
  contains_geom = NULL,
  typetup = NULL,
  fields = NULL,
  in_bbox = NULL,
  catpro3 = NULL
)

Arguments

token

token

code_insee

Code INSEE communal ou d'arrondissement municipal (il est possible d'en demander plusieurs (10 maximum), séparés par des virgules, dans le même département)

contains_geom

Renvoie les entités dont la géometrie contient celle précisée dans le filtre (en format GeoJSON, WKT, HEXEWKB, WKB). Exemple : /?contains_geom='type':'Point', 'coordinates':[2.17,46.75]

typetup

Type de l’entité : SIMPLE, PDLMP ou UF - cf typetup

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

in_bbox

Emprise rectangulaire (via Longitude min,Latitude min,Longitude max,Latitude max). L'emprise demandée ne doit pas excéder un carré de 0.02 deg. x 0.02 deg.

catpro3

Chaîne(s) de caractères contenue dans le code de catégorie de propriétaire (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : catpro3=P,F2,A2c - cf catpro3

Valeur de retour

Retourne, en GeoJSON, les tup issues des Fichiers Fonciers pour la commune ou l'emprise demandée (paramètre code_insee ou in_bbox obligatoire)

Exemples

if (FALSE) {
ff_geotups(code_insee=59001,token=token)
}