Locaux

Locaux

Utilisation

ff_locaux(
  token,
  code_insee = NULL,
  dteloc = NULL,
  fields = NULL,
  idpar = NULL,
  idprocpte = NULL,
  idsec = NULL,
  locprop = NULL,
  loghlls = NULL,
  ordering = NULL,
  proba_rprs = NULL,
  slocal_max = NULL,
  slocal_min = NULL,
  typeact = NULL,
  catpro3 = NULL
)

Arguments

token

token

code_insee

Code INSEE communal ou d'arrondissement municipal (possibilité de passer un vecteur de code insee sans limite maximum)

dteloc

Type(s) de local (il est possible de spécifier plusieurs types et de séparer par une virgule) - exemple : dteloc=1,2 - cf dteloc

fields

Retourne tous les champs associés si fields=all, sinon retourne uniquement une selection de champs.

idpar

Identifiant de parcelle - cf idpar

idprocpte

Identifiant de compte communal - cf idprocpte

idsec

Identifiant de section cadastrale - cf idsec

locprop

Localisation généralisée du propriétaire recevant la Taxe Foncière - cf locprop

loghlls

Logement d’habitation de type logement social repéré par exonération - cf loghlls

ordering

Which field to use when ordering the results.

proba_rprs

Probabilité de résidence principale ou secondaire (il est possible de spécifier plusieurs types et de séparer par une virgule) - exemple : proba_rprs=AC,NO - cf proba_rprs

slocal_max

Surface maximale des parties d'évaluation (m2) - cf slocal

slocal_min

Surface minimale des parties d'évaluation (m2) - cf slocal

typeact

Chaîne(s) de caractères contenue dans le classement du local selon le type d'activité (Code catégorie du local d’activité) (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : typeact=MAG,SPE3 - cf typeact

catpro3

Chaîne(s) de caractères contenue dans le code de catégorie de propriétaire (il est possible de ne specifier que les premiers niveaux et de séparer par une virgule) - exemple : catpro3=P,F2,A2c - cf catpro3

Valeur de retour

Retourne les locaux issus des Fichiers Fonciers pour la commune demandée (paramètre code_insee obligatoire)

Exemples

if (FALSE) {
ff_locaux(code_insee='59001',token=token)
}